Place tests for each API endpoint into a separate file
