+++ server/test/api/shared.ts
@@ -1,0 +1,2 @@
+
+export const SHOWCASE_TABLE = 'datatypeshowcase';
